cambios de id
(cherry picked from commit b3cf32406305690e304e39a43d019b1bdfedff58)

+++ src/main/java/com/bew/demo/dao/DictamenRepository.java
@@ -15,7 +15,7 @@ public interface DictamenRepository  extends JpaRepository <Dictamen,Integer>{
 	@Query(
             value = "SELECT s FROM Dictamen s WHERE s.idAlumno = :idAlumno",
             nativeQuery = false)
-    Optional<Dictamen> findByIdAlumno(@Param("idAlumno") Integer idAlumno);
+    Optional<Dictamen> findByIdAlumno(@Param("idAlumno") Long idAlumno);
 	
 	@Query(
             value = "SELECT s FROM Dictamen s WHERE s.estado = :estado",
